Minor improvements to Film Grain tool
I had intended to convert this to the Gaussian approximation blur, but
because the softening must be very gradual, the full Gaussian function
is required.

So instead I made some minor performance and quality tweaks to the tool.

+++ Modules/VBP_ProcessorModule.bas
@@ -1035,7 +1035,7 @@ Public Sub Process(ByVal processID As String, Optional showDialog As Boolean = F
             If showDialog Then
                 showPDDialog vbModal, FormFilmGrain
             Else
-                FormFilmGrain.AddFilmGrain cParams.GetDouble(1), cParams.GetLong(2)
+                FormFilmGrain.AddFilmGrain cParams.GetDouble(1), cParams.GetDouble(2)
             End If
         
         Case "Add RGB noise"
